+++ mu3d/gluetab.cpp
@@ -20,23 +20,23 @@ namespace mu3d
 		}
 
 		// get the bottom left corner of the base
-		_bl = glm::vec3(float(_edge._halfedge->vertex()->point().x()),
-			float(_edge._halfedge->vertex()->point().y()),
-			float(_edge._halfedge->vertex()->point().z()));
+		_bl = glm::dvec3(_edge._halfedge->vertex()->point().x(),
+			_edge._halfedge->vertex()->point().y(),
+			_edge._halfedge->vertex()->point().z());
 
 		// get the bottom right corner of the base
-		_br = glm::vec3(float(_edge._halfedge->prev()->vertex()->point().x()),
-			float(_edge._halfedge->prev()->vertex()->point().y()),
-			float(_edge._halfedge->prev()->vertex()->point().z()));
+		_br = glm::dvec3(_edge._halfedge->prev()->vertex()->point().x(),
+			_edge._halfedge->prev()->vertex()->point().y(),
+			_edge._halfedge->prev()->vertex()->point().z());
 
 		// get the vertex where the gluetab is extruded to
-		glm::vec3 target;
+		glm::dvec3 target;
 		Polyhedron::Halfedge_around_facet_circulator hfc = face->facet_begin();
 		do
 		{
-			target = glm::vec3(float(hfc->vertex()->point().x()),
-				float(hfc->vertex()->point().y()),
-				float(hfc->vertex()->point().z()));
+			target = glm::dvec3(hfc->vertex()->point().x(),
+				hfc->vertex()->point().y(),
+				hfc->vertex()->point().z());
 
 			// if the vertex is neither the bottom right or bottom left it is the target
 			if (target != _bl && target != _br)
@@ -46,10 +46,10 @@ namespace mu3d
 		} while (++hfc != face->facet_begin());
 
 
-		glm::vec3 side = (_br - _bl) / 8.0f;
+		glm::dvec3 side = (_br - _bl) / 8.0;
 
-		_tr = _br + (target - _br) / 3.0f - side * 1.8f;
-		_tl = _bl + (target - _bl) / 3.0f + side * 1.8f;
+		_tr = _br + (target - _br) / 3.0 - side * 1.8;
+		_tl = _bl + (target - _bl) / 3.0 + side * 1.8;
 
 		_br = _br - side;
 		_bl = _bl + side;
