+++ mu3d/mu3d/facetoplane.h
@@ -11,26 +11,26 @@ namespace mu3d
         faceToPlane();
         ~faceToPlane();
 
-        glm::vec3 A;
-        glm::vec3 B;
-        glm::vec3 C;
+        glm::dvec3 A;
+        glm::dvec3 B;
+        glm::dvec3 C;
 
-        glm::vec2 a;
-        glm::vec2 b;
-        glm::vec2 c;
+        glm::dvec2 a;
+        glm::dvec2 b;
+        glm::dvec2 c;
 
         bool _overlaps;
 
         long parent;
         long self;
 
-        glm::vec2 const& get(glm::vec3 const& vec);
-        glm::vec3 const& get(glm::vec2 const& vec);
-        glm::vec3 const& get(glm::vec3 const& one, glm::vec3 const& two);
+        glm::dvec2 const& get(glm::dvec3 const& vec);
+        glm::dvec3 const& get(glm::dvec2 const& vec);
+        glm::dvec3 const& get(glm::dvec3 const& one, glm::dvec3 const& two);
 
         double overlaps(faceToPlane& other);
     private:
-        static glm::vec3 _color;
-        static glm::vec3 _colorOverlap;
+        static glm::dvec3 _color;
+        static glm::dvec3 _colorOverlap;
     };
 }
